Add timeout option (#26)
* Add timeout option for requests

Based on an old PR at https://github.com/cjoudrey/gluahttp/pull/20 that
seems to have been deleted.

The timeout option can be set to a number, which will be cast to an int
and interpreted as seconds, or a string which will be parsed by time.ParseDuration.

Add tests, one case should timeout and one that shouldn't.

* Update readme with info about timeout option

* Use WithContext() instead of Clone()

Clone is not supported in older versions of Go
and WithContext seems to work just as well.

* Catch duration parsing errors, clean up imports

Co-authored-by: Johannes Larsson <j@johannes-larsson.se>